requirements.txt
Flask
firebase-admin
requests
python-dotenv
Flask-Cors
Werkzeug
gunicorn
